+++ data-raw/df_ASOIAF.R
@@ -8,7 +8,7 @@ library(BGmisc)
 
 
 ## Create dataframe
-ASOIAF <- readGedcom("data-raw/ASOIAF.ged")  %>%
+ASOIAF <- readGedcom("data-raw/ASOIAF.ged") %>%
   mutate(name = str_remove(name, "/"))
 
 # ASOIAF <- readGedcom("data-raw/ASOIAF_040725.ged")
@@ -19,7 +19,7 @@ df <- ped2fam(ASOIAF, personID = "personID") %>%
     -name_nsfx,
     -name_given,
     -name_surn,
-#    -name_marriedsurn,
+    #    -name_marriedsurn,
     -death_caus,
     -FAMC,
     -FAMS
@@ -27,41 +27,36 @@ df <- ped2fam(ASOIAF, personID = "personID") %>%
   mutate(
     momID = as.numeric(momID),
     dadID = as.numeric(dadID),
-    name = case_when(name=="Naerys " ~ "Naerys Targaryen",
-                     name=="Rhaenyra " ~ "Rhaenyra Targaryen",
-                     name == "Betharios " ~ "Betharios of Braavos",
-                              personID == 257 ~ "Princess Of Dorne",
-                              name=="Rowena " ~ "Rowena Arryn",
-                              name=="Pate " ~ "Pate of the Blue Fork",
-                              name=="Mellario " ~ "Mellario of Norvos",
-                              TRUE ~ name),
+    name = case_when(
+      name == "Naerys " ~ "Naerys Targaryen",
+      name == "Rhaenyra " ~ "Rhaenyra Targaryen",
+      name == "Betharios " ~ "Betharios of Braavos",
+      personID == 257 ~ "Princess Of Dorne",
+      name == "Rowena " ~ "Rowena Arryn",
+      name == "Pate " ~ "Pate of the Blue Fork",
+      name == "Mellario " ~ "Mellario of Norvos",
+      TRUE ~ name
+    ),
     twinID = case_match(name,
       "Jaime Lannister" ~ 165,
       "Cersei Lannister" ~ 164,
-
       "Alyn Frey" ~ 73,
       "Androw Frey" ~ 74,
-
-      "Aegon Blackfyre"~ 314,
+      "Aegon Blackfyre" ~ 314,
       "Aemon Blackfyre" ~ 313,
-
       "Dickon Frey" ~ 117,
       "Mathis Frey" ~ 116,
-
       "Jaime Frey" ~ 101,
       "Tywin Frey" ~ 102,
-
-      "Sarra Frey" ~99,
+      "Sarra Frey" ~ 99,
       "Serra Frey" ~ 98,
-
       "Martyn Lannister" ~ 173,
       "Willem Lannister" ~ 174,
-
       "Hobber 'Slobber' Redwyne" ~ 390,
       "Horas 'Horror' Redwyne" ~ 391,
-
       .default = NA_real_
-    ))
+    )
+  )
 
 # pedADD <- ped2com(df , personID = "id", momID = "momID",
 # dadID = "dadID", component = "additive", isChild_method = "partial_parent")