+++ tests/testthat/test-tweakPedigree.R
@@ -24,19 +24,25 @@ test_that("makeTwins - Twins specified by IDs", {
   result <- makeTwins(ped, ID_twin1 = 1, ID_twin2 = 2)
 
   expect_equal(result, expected_result)
+
   # does it handle weird variable names? "fam" = "famID"
 
   names(ped)[1] <- "fam"
 
-  result <- makeTwins(ped, ID_twin1 = 1, ID_twin2 = 2, verbose = TRUE)
-  expect_equal(result, expected_result)
+  result_badfam <- makeTwins(ped, ID_twin1 = 1, ID_twin2 = 2, verbose = TRUE)
 
-  result2 <- makeTwins(ped, ID_twin1 = 1,
-                       ID_twin2 = 2,
-                       verbose = TRUE, zygosity = "DZ")
-  expected_result$zygosity[expected_result$zygosity=="MZ"] <- "DZ"
+  expect_equal(result_badfam, expected_result)
+
+
+  result2 <- makeTwins(ped,
+    ID_twin1 = 1,
+    ID_twin2 = 2,
+    verbose = TRUE, zygosity = "DZ"
+  )
+
+  expected_result$zygosity[expected_result$zygosity == "MZ"] <- "DZ"
   expect_equal(result2, expected_result)
- # hp <- makeTwins(potter, ID_twin1 = 12, ID_twin2 = 13, verbose = TRUE)
+  # hp <- makeTwins(potter, ID_twin1 = 12, ID_twin2 = 13, verbose = TRUE)
 })
 
 test_that("makeTwins - Twins specified by generation", {
@@ -64,6 +70,24 @@ test_that("makeTwins - Twins specified by generation", {
   expect_equal(length(unique(result$momID[!is.na(result$twinID)])), 1)
   # do they have the same dad?
   expect_equal(length(unique(result$dadID[!is.na(result$twinID)])), 1)
+
+  resultdz <- makeTwins(ped, gen_twin = gen_twin, zygosity = "DZ")
+
+  expect_equal(names(resultdz), c("famID", "ID", "gen", "dadID", "momID", "spID", "sex", "twinID", "zygosity"))
+  # do we have the same people?
+  expect_equal(resultdz$ID, ped$ID)
+  # did it make one pair of twins?
+  expect_equal(sum(!is.na(resultdz$twinID)), 2)
+  # did it make the pair in the correct generation?
+  expect_equal(mean(resultdz$gen[!is.na(resultdz$twinID)]), gen_twin)
+  # are they the same sex?
+  expect_equal(length(unique(resultdz$sex[!is.na(resultdz$twinID)])), 1)
+  # are they from the same family?
+  expect_equal(length(unique(resultdz$fam[!is.na(resultdz$twinID)])), 1)
+  # do they have the same mom?
+  expect_equal(length(unique(resultdz$momID[!is.na(resultdz$twinID)])), 1)
+  # do they have the same dad?
+  expect_equal(length(unique(resultdz$dadID[!is.na(resultdz$twinID)])), 1)
 })
 
 # Test for makeInbreeding function